Move PartitionedManagedLedgerInfo to broker module
